+++ Touch/TouchSimulate.h
@@ -1,0 +1,14 @@
+//
+//  TouchSimulate.h
+//  LinearMouse
+//
+//  Created by lujjjh on 2021/8/5.
+//
+
+#ifndef TouchSimulate_h
+#define TouchSimulate_h
+
+void simulateSwipeLeft(void);
+void simulateSwipeRight(void);
+
+#endif /* TouchSimulate_h */